% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{annotate.electrophoresis}
\alias{annotate.electrophoresis}
\title{Add annotations to an electrophoresis object}
\usage{
annotate.electrophoresis(
  electrophoresis,
  annotations,
  header = TRUE,
  row.names = NULL,
  sep = "\\t",
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{...}{Additional arguments passed to \code{\link{read.table}}.}

\item{annotation}{Either a data frame or the path of a file that can be read by \code{\link{read.table}}.}
}
\value{
A new \code{electrophoresis} object with additional columns added to its \code{$samples} element.
}
\description{
This function adds columns of annotations to the \code{$samples} table of an \code{electrophoresis} object. These new annotations can be used for subsetting, plot faceting, etc.
}
\details{
The input annotations can be a data frame or either a \code{\link{connection}} or a \code{\link{character}} containing the path of a file, which is then read into a data frame by \code{\link{read.table}}. Therefore a file will probably need to be in CSV format or similar.

The first column of \code{annotations} is used to identify samples in the matching column of \code{electrophoresis$samples}, so it needs a matching column name, usually \code{sample.name} or \code{well.number}. But you can use any column that exists in the table, such as \code{batch} or \code{reagent.id} or even a column added by a previous call to this function. The elements of this first column are interpreted as class \code{\link{character}} and cannot contain any duplicates. However, this column does not need to be in the same order as \code{electrophoresis$samples} and not all identifiers in the sample table (e.g. sample names) need to appear in the annotation table, nor vice versa.

Each additional column of \code{annotations} produces a new column of the same name in \code{electrophoresis$samples}. If there is already a column of that name, it is overwritten. Samples that do not appear in the annotation table receive \code{NA} for their new annotations. Samples with the same identifier (e.g. the same \code{sample.name} because they are replicates) receive the same annotations.
}
