% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{between.markers}
\alias{between.markers}
\title{Check whether data points are between markers}
\usage{
between.markers(electrophoresis)
}
\arguments{
\item{electrophoresis}{}
}
\value{
A vector of logicals with length \code{nrow{electrophoresis$data}}.
}
\description{
This function takes an electrophoresis object and reports whether each data point is between the lower and upper length markers.
}
\details{
Observations are considered to be between the markers if they are above the upper boundary of the lower marker and the lower boundary of the upper marker, as reported in the Agilent software's peak detection. If there is no upper marker, all points above the upper boundary of the lower marker are considered to be between the markers.
}
\seealso{
\code{\link{\in.peaks}}
}
