% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{dv200}
\alias{dv200}
\title{DV200 analysis}
\usage{
dv200(
  electrophoresis,
  prop.variable = "concentration",
  lower.marker.spread = 1
)
}
\arguments{
\item{electrophoresis}{An \code{\link{electrophoresis}} object.}

\item{prop.variable}{Which variable to use for the proportion.}

\item{lower.marker.spread}{Amount to scale the width of the lower marker peak.}
}
\description{
Calculate the proportion of fragments above 200 bases (DV200). Only fragments between the lower and upper markers are considered.
}
\details{
Note: Despite unclear wording, Agilent calculates DV200 as a proportion of total mass (concentration), rather than a proportion of molecules (molarity). For some purposes molar DV200 may be more relevant, but current protocols refer to the default DV200 based on mass.

By default \code{lower.marker.spread = 1} (see \code{\link{between.markers}}) for consistency with the Agilent software.
}
\seealso{
\code{\link{region.ratio}}, \code{\link{illumina.library.ratio}}
}
