% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{electrophoresis}
\alias{electrophoresis}
\title{Electrophoresis class}
\usage{
electrophoresis(
  data = NULL,
  assay.info = NULL,
  samples = NULL,
  peaks = NULL,
  regions = NULL,
  mobility.functions = NULL
)
}
\arguments{
\item{data}{A tall data frame of the run data, specifically:
\itemize{
\item \code{time} (Bioanalyzer, ProSize) - time when this data point was measured
\item \code{aligned.time} (Bioanalyzer, ProSize) - measurement time aligned between the expected times of the marker peaks
\item \code{distance} (TapeStation) - migration distance of the measurement from the top of the gel area
\item \code{relative.distance} (TapeStation) - migration distance normalized relative to the marker peaks
\item \code{fluorescence} - fluorescence reading at this point
\item \code{length} - estimated molecule length at this point
\item \code{concentration} - estimated concentration of the area under the curve between this point and the previous one
\item \code{molarity} - estimated molarity of the area under the curve between this point and the previous one
}}

\item{assay.info}{A list of metadata about each batch and the assay kit used.}

\item{samples}{A data frame of metadata for each sample (also annotated to \code{data}, \code{peaks}, and \code{regions} with the same factor levels), specifically:
\itemize{
\item \code{batch} - the batch (instrument run) of the sample, from the file name
\item \code{well.number} - the well number in which the sample was loaded
\item \code{sample.name} - the name of the sample
\item \code{sample.observations} - notes about this sample supplied by the user or the Agilent software
\item \code{sample.comment} - notes about this sample supplied by the user
\item \code{reagent.id} (TapeStation) - the name of the ScreenTape used for this sample
\item \code{ladder.well} - which well contains the ladder that calibrates this sample
\item \code{RIN}, \code{DIN}, \code{RQN}, \verb{28S/18S}, etc. - sample quality metrics reported by the Agilent software
}}

\item{peaks}{A data frame of peaks reported by the Agilent software, annotated with their lower and upper boundaries in various scales.}

\item{regions}{A data frame of regions of interest reported by the Agilent software, annotated with their lower and upper boundaries in varous scales.}

\item{mobility.functions}{A list of model functions, one per ladder used for calibration, to convert migration speed measurements (aligned time or relative distance) into estimated molecule lengths.}
}
\description{
This S3 class is a generic container for electrophoresis data and metadata. The constructor simply assembles the provided members into a class structure without checking them: if they are formatted incorrectly there will be no warnings at this stage.
}
