% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{from.sample}
\alias{from.sample}
\title{Check whether data points are from a certain sample}
\usage{
from.sample(electrophoresis, which.sample)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{which.sample}{The integer index of a sample in \code{electrophoresis$samples}.}
}
\value{
A vector of logicals with length \code{nrow(data)}.
}
\description{
This function takes an electrophoresis object and checks whether each data point is from a specified sample.
}
\seealso{
\code{\link{from.samples}}, \code{\link{in.peak}}, \code{\link{in.region}}
}
