% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{from.samples}
\alias{from.samples}
\title{Match data points to the samples they are from}
\usage{
from.samples(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}
}
\value{
A vector of integers with length \code{nrow(data)}. Each element is the integer index of the sample in \code{electrophoresis$samples} that the data point belongs to.
}
\description{
This function takes an electrophoresis object and reports which of the samples each data point belongs to.
}
\seealso{
\code{\link{from.sample}}, \code{\link{in.peaks}}, \code{\link{in.regions}}
}
