% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{get.x.name}
\alias{get.x.name}
\title{Get the original x-variable}
\usage{
get.x.name(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}
}
\value{
A character giving the name of the x-value.
}
\description{
This function takes an \code{electrophoresis} object and returns the name of the x-value that was used to fit the mobility model.
}
\details{
The result should only be either \code{"aligned time"} for Bioanalyzer data or \code{"relative.distance"} for TapeStation data.
}
