% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{illumina.library.ratio}
\alias{illumina.library.ratio}
\title{Ratio of good inserts to adapter dimers}
\usage{
illumina.library.ratio(
  electrophoresis,
  min.sequenceable = 100,
  min.good.insert = 200,
  max.sequenceable = 700
)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{min.sequenceable}{The shortest length of molecules that are likely to produce data on the sequencer (ignore shorter molecules that are probably just unclusterable free primers or adapters).}

\item{min.good.insert}{The shortest length of a desirable molecule (adapter length plus insert length).}

\item{max.sequenceable}{The longest length of molecules that are likely to produce data on the sequencer (ignore molecules that are too long for clustering).}
}
\description{
For Illumina sequencing libraries, compute the molar ratio of molecules with desirably long inserts to undesirable adapter dimers in each sample.
}
