% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.custom.region}
\alias{in.custom.region}
\title{Check whether data points are within a custom region}
\usage{
in.custom.region(
  data,
  lower.bound = -Inf,
  upper.bound = Inf,
  bound.variable = "length"
)
}
\arguments{
\item{data}{A data frame of electrophoresis data, from the \code{$data} member of an \code{electrophoresis} object (not the whole object itself).}

\item{lower.bound}{Lower boundary of the region.}

\item{upper.bound}{Upper boundary of the region.}

\item{bound.variable}{Which variable the boundaries refer to.}
}
\value{
A vector of logicals with length \code{nrow(data)}.
}
\description{
This function takes the \code{$data} element of an \code{electrophoresis} object and the boundaries of a region, for a desired variable, and checks whether each data point is within that region.
}
\seealso{
\code{\link{in.region}}, \code{\link{in.peak}}
}
