% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.peak}
\alias{in.peak}
\title{Check whether data points are within a reported peak}
\usage{
in.peak(electrophoresis, which.peak)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{which.peak}{An integer indicating one of the peaks in \code{electrophoresis$peaks}.}
}
\value{
A vector of logicals with length \code{nrow(data)}.
}
\description{
This function takes an electrophoresis object and checks whether each data point is within a specified peak.
}
\details{
Because each peak in the table belongs to a specific sample, only data points from that sample are \code{TRUE}.
}
\seealso{
\code{\link{in.region}}, \code{\link{in.custom.region}}, \code{\link{in.peaks}}
}
