% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.peak.region}
\alias{in.peak.region}
\alias{in.peak}
\alias{in.region}
\title{Check whether data points are within a reported peak or region}
\usage{
in.peak(electrophoresis, which.peak)

in.region(electrophoresis, which.region)
}
\arguments{
\item{electrophoresis}{An \code{\link{electrophoresis}} object.}

\item{which.peak}{The integer index of a peak in \code{electrophoresis$peaks}.}

\item{which.region}{The integer index of a region in \code{electrophoresis$regions}.}
}
\value{
A vector of logicals with length \code{nrow(electrophoresis$data)}.
}
\description{
These functions take an electrophoresis object and check whether each data point is within a specified peak or region.
}
\details{
Data points are considered to be within the boundaries of a peak if their original x-value (`aligned.time` for Bioanalyzer, `relative.distance` for TapeStation) is within the boundaries reported by the Agilent software, but for regions the length boundaries are used.
}
\seealso{
\code{\link{in.custom.region}}, \code{\link{in.peaks}}, \code{\link{in.regions}}
}
