% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.peaks}
\alias{in.peaks}
\title{Match data points to the peaks they are in}
\usage{
in.peaks(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}
}
\value{
A vector of integers with length \code{nrow(data)}. Each element is either the integer index of the peak in \code{electrophoresis$peaks} that the data point belongs to, or NA if it is not in any of the annotated peaks.
}
\description{
This function takes an electrophoresis object and reports which of the peaks each data point belongs to, if any.
}
\details{
Warning: If peaks in the reported table overlap, any data point in more than one peak will only be matched to the last peak it belongs to.
}
\seealso{
\code{\link{in.regions}}, \code{\link{in.peak}}, \code{\link{between.markers}}
}
