% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.peaks.regions}
\alias{in.peaks.regions}
\alias{in.peaks}
\alias{in.regions}
\title{Match data points to the peaks or regions they are in}
\usage{
in.peaks(electrophoresis)

in.regions(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}
}
\value{
A vector of integers with length \code{nrow(data)}. Each element is either the integer index of the peak in \code{electrophoresis$peaks} or region in \code{electrophoresis$regions} that the data point belongs to, or NA if it is not in any of the annotated peaks or regions.
}
\description{
These functions take an electrophoresis object and report which of the peaks or regions each data point belongs to, if any.
}
\details{
Warning: If peaks or regions in the reported table overlap, any data point in more than one peak or region will only be matched to the last peak or region it belongs to.
}
\seealso{
\code{\link{in.peak}}, \code{\link{in.region}}, \code{\link{between.markers}}
}
