% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.region}
\alias{in.region}
\title{Check whether data points are within a reported region}
\usage{
in.region(electrophoresis, which.region)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{which.region}{The integer index of a region in \code{electrophoresis$regions}.}
}
\value{
A vector of logicals with length \code{nrow(electrophoresis$data)}.
}
\description{
This function takes an electrophoresis object and checks whether each data point is within a specified region.
}
\details{
Because each region in the table belongs to a specific sample, only data points from that sample are \code{TRUE}.
}
\seealso{
\code{\link{in.peak}}, \code{\link{in.custom.region}}, \code{\link{in.regions}}
}
