% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{in.regions}
\alias{in.regions}
\title{Match data points to the regions they are in}
\usage{
in.regions(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}
}
\value{
A vector of integers with length \code{nrow(data)}. Each element is either the integer index of the region in \code{electrophoresis$regions} that the data point belongs to, or NA if it is not in any of the annotated regions.
}
\description{
This function takes an electrophoresis object and reports which of the regions each data point belongs to, if any.
}
\details{
Warning: If regions in the reported table overlap, any data point in more than one region will only be matched to the last region it belongs to.
}
\seealso{
\code{\link{in.peaks}}, \code{\link{in.region}}
}
