% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{integrate.custom}
\alias{integrate.custom}
\title{Integrate a variable in a custom region}
\usage{
integrate.custom(electrophoresis, ...)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{...}{Arguments passed to \code{\link{integrate.rawdata}}. In particular use \code{lower.bound} and \code{upper.bound} to set the boundaries, \code{bound.variable} for the boundary variable, and \code{sum.variable} for the variable to sum.}
}
\description{
Compute the sum of some electrophoresis variable in an \code{electrophoresis} object between specified boundaries. The variable is summed individually for each sample.
}
\seealso{
\code{\link{integrate.rawdata}}, \code{\link{integrate.peaks}}, \code{\link{integrate.regions}}, \code{\link{region.ratio}}
}
