% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.peak.region}
\alias{integrate.peak.region}
\alias{integrate.peak}
\alias{integrate.region}
\title{Integrate a variable in peaks or regions}
\usage{
integrate.peak(electrophoresis, index, sum.variable = "concentration")

integrate.region(electrophoresis, index, sum.variable = "concentration")
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{index}{The index, or a vector of indexes, of the peaks or regions to integrate (row numbers in \code{electrophoresis$peaks} or \code{electrophoresis$regions}).}

\item{sum.variable}{Which variable to sum in each peak.}
}
\description{
Compute the sum of some electrophoresis variable between the boundaries of reported peaks or regions in an \code{electrophoresis} object.
}
\seealso{
\code{\link{integrate.custom}}
}
