% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.peaks}
\alias{integrate.peaks}
\title{Integrate a variable under each peak}
\usage{
integrate.peaks(electrophoresis, sum.variable = "molarity")
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{sum.variable}{Which variable to sum in each peak.}
}
\description{
Compute the sum of some electrophoresis variable between the boundaries of each reported peak in an \code{electrophoresis} object.
}
\seealso{
\code{\link{integrate.rawdata}}, \code{\link{integrate.regions}}, \code{\link{integrate.custom}}
}
