% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.peaks.regions}
\alias{integrate.peaks.regions}
\alias{integrate.peaks}
\alias{integrate.regions}
\title{Integrate a variable in each peak or region}
\usage{
integrate.peaks(electrophoresis, sum.variable = "concentration")

integrate.regions(electrophoresis, sum.variable = "concentration")
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{sum.variable}{Which variable to sum in each peak.}
}
\description{
Compute the sum of some electrophoresis variable between the boundaries of each reported peak or region in an \code{electrophoresis} object.
}
\seealso{
\code{\link{integrate.custom}}
}
