% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{integrate.rawdata}
\alias{integrate.rawdata}
\title{Integrate a variable under the electrophoresis curve}
\usage{
integrate.rawdata(
  data,
  lower.bound = NULL,
  upper.bound = NULL,
  bound.variable = "length",
  peak = NULL,
  sum.variable = "molarity"
)
}
\arguments{
\item{data}{A data frame of electrophoresis data, from the \code{$data} member of an electrophoresis object (not the whole object itself).}

\item{lower.bound}{Lower boundary of the region to integrate, or \code{NULL} to extend to negative infinity.}

\item{upper.bound}{Upper boundary of the region to integrate, or \code{NULL} to extend to positive infinity.}

\item{bound.variable}{Which variable the boundaries refer to, e.g. \code{"length"}.}

\item{peak}{The index of the peak to integrate in, or \code{NULL} to ignore peaks.}

\item{sum.variable}{Which variable to sum in the target region.}
}
\description{
Compute the sum of some electrophoresis variable between some specified boundaries, or within an annotated peak, or both (intersection).
}
\details{
This is an all-purpose base function that accepts any combination of sum variable, boundaries, boundary variable, and peak. Crucially, it operates directly on the \code{$data} member of an \code{electrophoresis} object rather than the entire object, and does not separate the results by sample. You can achieve this manually by taking a \code{\link{subset}} of the data but it is probably easier to use one of the functions that wrap this one.
}
\seealso{
\code{\link{integrate.peaks}}, \code{\link{integrate.regions}}, \code{\link{integrate.custom}}
}
