% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.regions}
\alias{integrate.regions}
\title{Integrate a variable in each region}
\usage{
integrate.regions(electrophoresis, sum.variable = "molarity")
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{sum.variable}{Which variable to sum in each region.}
}
\description{
Compute the sum of some electrophoresis variable between the boundaries of each region of interest in an \code{electrophoresis} object.
}
\seealso{
\code{\link{integrate.rawdata}}, \code{\link{integrate.peaks}}, \code{\link{integrate.custom}}
}
