% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{labeller.electrophoresis}
\alias{labeller.electrophoresis}
\title{Labeller for electrophoresis samples}
\usage{
labeller.electrophoresis(electrophoresis)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object whose \code{samples} member contains all the samples in your plot.}
}
\description{
This is a labeller function compatible with \code{\link{facet_wrap}} and \code{\link{facet_grid}}. It allows you to facet the data from an \code{electrophoresis} object on \code{sample.index}, to keep the samples in the observed order, but replaces the facet labels with the annotated sample names.
}
