% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/molecular_weight.R
\name{molecular.weight}
\alias{molecular.weight}
\title{Estimate molecular weight of nucleic acids by length}
\usage{
molecular.weight(length, type)
}
\arguments{
\item{length}{The length of the nucleic acid in bases (nt for RNA, bp for DNA).}

\item{type}{The type of nucleic acid: \code{"RNA"} for single-stranded RNA, \code{"DNA"} for double-stranded DNA.}
}
\value{
The estimated molecular weight in daltons.
}
\description{
This function estimates the molecular weight of a single-stranded RNA or double-stranded DNA from its length in bases.
}
\references{
Thermo Fisher Scientific: DNA and RNA Molecular Weights and Conversions. Accessed 13 March 2020. \url{https://www.thermofisher.com/us/en/home/references/ambion-tech-support/rna-tools-and-calculators/dna-and-rna-molecular-weights-and-conversions.html}
}
