% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{normalize.proportion}
\alias{normalize.proportion}
\title{Normalize data to proportions}
\usage{
normalize.proportion(electrophoresis, variable)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{variable}{The name of a variable in \code{electrophoresis$data}.}
}
\value{
A vector whose length equals \code{nrow(electrophoresis$data)}, in which each value is an observation of the given variable normalized by the total for that sample, or NA if the observation is not between the markers.
}
\description{
For a given variable, normalize the data values in an \code{electrophoresis} object to proportions of the total. After normalization the sum of the variable for each sample is 1.
}
\details{
Only data points between the markers are considered in the total. If there is no upper marker, all data points above the lower marker are considered.
}
\seealso{
\code{\link{scale.by.differential}}
}
