% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{qc.electrophoresis}
\alias{qc.electrophoresis}
\title{Compare estimates with Agilent software's output}
\usage{
qc.electrophoresis(electrophoresis, variable, log = TRUE)
}
\arguments{
\item{electrophoresis}{An \code{\link{electrophoresis}} object.}

\item{variable}{Which variable to compare. One of \code{"length"}, \code{"concentration"}, \code{"molarity"}.}

\item{log}{Whether to log-scale both axes. Typically the data are more evenly spaced in log scale.}
}
\value{
A ggplot object containing several layers. You can draw it directly or customize it like any other ggplot object by adding more layers.
}
\description{
This function is a shortcut to plot the estimates of a variable from the \code{\link{bioanalyzeR}} package against the values reported by the Agilent software, at all reported peaks. This allows you to check the quality of the estimates.
}
\details{
If \code{variable = "length"}, apply the mobility model(s) fit by \code{\link{read.bioanalyzer}} or \code{\link{read.tapestation}} to the peak centers' aligned times or relative distance and compare those estimated molecule lengths with the lengths reported by the Agilent software. If \code{variable = "concentration"} or \code{variable = "molarity"}, integrate the appropriate variable between the boundaries of each peak and compare that sum with the figure reported by the Agilent software. For the ladder Agilent's reported values are hardcoded to the known, correct properties of the ladder molecules.
}
\seealso{
\code{\link{stdcrv.mobility}}
}
