% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{qplot.electrophoresis}
\alias{qplot.electrophoresis}
\title{Plot electrophoresis data}
\usage{
qplot.electrophoresis(
  electrophoresis,
  x = "length",
  y = "molarity",
  log = "",
  facets = ~sample.index,
  margins = FALSE,
  scales = "fixed",
  geom = "line",
  include.ladder = FALSE,
  between.markers = TRUE,
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  peak.fill = "darkred",
  region.alpha = 0.2,
  area.alpha = 0.2,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{x}{The variable to use as the x-value of each point in the graph. Can be one of \code{"time"}, \code{"aligned.time"}, \code{"distance"}, \code{"relative.distance"}, or \code{"length"}.}

\item{y}{The variable to use as the y-value of each point in the graph. Can be one of \code{"fluorescence"}, \code{"concentration"}, or \code{"molarity"}.}

\item{log}{Which variables to log-transform (\code{"x"}, \code{"y"}, or \code{"xy"}).}

\item{facets}{Faceting formula to use. Picks \code{\link{facet_wrap}} or \code{\link{facet_grid}} depending on whether the formula is one- or two-sided. If \code{NULL}, overlay all samples in one color-coded graph.}

\item{margins}{Display marginal facets (via \code{\link{facet_grid}}) if using a two-side faceting formula.}

\item{scales}{Scaling rules for the facets, passed to \code{\link{facet_wrap}}.}

\item{geom}{Name of the geom to draw. Currently only \code{"line"} (\code{\link{geom_line}}, to get continuous lines) and \code{"area"} (\code{\link{geom_area}}, to fill the area under the curves) are supported.}

\item{include.ladder}{If \code{FALSE}, graph only the actual samples and not the ladder(s) wells.}

\item{between.markers}{If \code{TRUE}, graph only data between the marker peaks.}

\item{xlim, }{ylim Limits of x- and y-axes}

\item{peak.fill}{Color to fill the area under reported peaks. Set to \code{NA} to skip plotting the peaks.}

\item{region.alpha}{Alpha-transparency of the highlight in the reported regions of interested. Set to \code{NA} to skip plotting the regions.}

\item{area.alpha}{Alpha-transparency of the filled areas under the curves, if they are overlaid in one graph (\code{facet = FALSE} and \code{geom = "area"}), to make them visible through one another.}

\item{title, }{xlab, ylab Plot title, x-axis label, and y-axis label.}
}
\value{
A ggplot object containing several layers. You can draw it directly or customize it like any other ggplot object by adding more layers.
}
\description{
This function is a shortcut to plot the data from an \code{electrophoresis} object, wrapping \code{\link{ggplot}} similarly to \code{\link{qplot}}. The result is analogous to electropherograms produced by the Agilent software.
}
\details{
Before plotting, the y-variable is scaled by the differentials in the x-value. Thus the units of the y-axis are divided by the units of the x-axis, e.g. molarity per length. This ensures that the area under the curve (width times height) represents the desired variable in the correct units. For example, if the x-variable is length in bp, the graph will be equivalent to a histogram with one bar for each possible molecule length in base pairs.
}
