% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{rawplot.electrophoresis}
\alias{rawplot.electrophoresis}
\title{Plots of raw electrophoresis data}
\usage{
rawplot.electrophoresis(electrophoresis, include.markers = TRUE, ...)
}
\description{
This function is a shortcut that calls \code{\link{qplot.electrophoresis}} with customized settings to show the raw data analogously to the Agilent software.
}
