% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{rbind.electrophoresis}
\alias{rbind.electrophoresis}
\title{Combine multiple electrophoresis objects}
\usage{
\method{rbind}{electrophoresis}(...)
}
\arguments{
\item{...}{Two or more objects of class \code{electrophoresis}.}
}
\value{
A new \code{electrophoresis} object containing all the data from the previous ones in the provided order.
}
\description{
This function combines multiple \code{electrophoresis} objects into one so you can analyze and graph multiple batches together.
}
\details{
All data frames are combined by \code{\link{rbind}} and lists are combined by \code{\link{c}}. Factor levels are expanded to the union of all inputs.
}
