% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioanalyzer.R, R/electrophoresis.R,
%   R/prosize.R, R/tapestation.R
\name{read.bioanalyzer}
\alias{read.bioanalyzer}
\alias{read.electrophoresis}
\alias{read.prosize}
\alias{read.tapestation}
\title{Read files into an electrophoresis object}
\usage{
read.bioanalyzer(xml.file, method = "hyman")

read.electrophoresis(
  ...,
  method = "hyman",
  mc.cores = if (.Platform$OS.type == "windows") 1 else detectCores()
)

read.prosize(csv.file, method = "hyman")

read.tapestation(xml.file, gel.image.file = NULL, method = "hyman")
}
\arguments{
\item{xml.file}{The filename of an XML file exported from the Bioanalyzer or TapeStation software. The XML file may be compressed with `gzip` and the filename can be a remote URL. The filename is expected to end in \code{.xml} or \code{.xml.gz} and the name before that extension is used as the name of the batch. TapeStation XML files must have corresponding PNG files with matching names.}

\item{method}{The method used to fit the mobility model of molecule length vs. migration distance, either \code{"interpolation"} (linear interpolation via \code{\link{approxfun}}), \code{"regression"} (log-linear regression via \code{\link{lm}} with the model \code{relative.distance ~ log(length)}), or one of the methods of \code{\link{splinefun}} (monotone methods recommended).}

\item{...}{One or more XML or electropherogram CSV files exported from the Bioanalyzer, TapeStation, or ProSize software.}

\item{mc.cores}{Maximum number of CPU cores to use (passed to \code{\link[parallel]{mclapply}}). Only one core is used per input file.}

\item{csv.file}{The filename of an electropherogram CSV file exported from the ProSize software. The filename must be in the original exported format and it must be accompanied by matching CSVs for the size calibration and peak table. The electropherogram must give the migration time, not estimated molecule size. The peak table must be exported in the "alternate" format and must contain the "From" and "To" columns. If a smear analysis CSV is present it will also be read.}

\item{gel.image.file}{The filename of a TapeStation gel image with blue highlight, in PNG format. The filename can be a URL.}
}
\description{
These functions read one or more XML or CSV files exported from the Agilent software (and accompanying PNG files if from a TapeStation) and calls the appropriate function to read them into an \code{electrophoresis} object, which is filled out with estimates of molecule length, concentration, and molarity. \code{read.electrophoresis} is the easiest to use as it automatically infers the correct file type.
}
\details{
Spline fitting seems to perform reasonably well on all data. Agilent appears to use linear interpolation with DNA data and log-linear regression on RNA data, so you could choose those options if you want to reproduce the results of the software more precisely. However, linear interpolation creates sudden spikes in the derivative that make the concentration and molarity estimates unstable; spline fitting is basically a smoother version of that. Log-linear regression is the standard theoretical approach but does not actually fit the data very well; more sophisticated parametric models may be added in the future.
}
\section{Functions}{
\itemize{
\item \code{read.bioanalyzer}: Read a Bioanalyzer XML file

\item \code{read.prosize}: Read ProSize CSV files

\item \code{read.tapestation}: Read a TapeStation XML and PNG file pair
}}

