% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosize.R
\name{read.prosize.electropherogram}
\alias{read.prosize.electropherogram}
\title{Read a ProSize electropherogram}
\usage{
read.prosize.electropherogram(csv.file)
}
\arguments{
\item{csv.file}{The filename of an electropherogram CSV exported by ProSize. The filename can be a URL.}
}
\value{
A list containing a data frame of the raw fluorescence data and a data frame of the sample metadata (a partial \code{electrophoresis} object).
}
\description{
This function reads an electropherogram (fluorescence vs. migration table) from the ProSize software saved in CSV format. The electropherogram must have been exported with migration times, not estimated sizes.
}
\seealso{
\code{\link{read.prosize}}, \code{\link{read.prosize.peaks}}, \code{\link{read.prosize.regions}}
}
