% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosize.R
\name{read.prosize.peaks}
\alias{read.prosize.peaks}
\title{Read a ProSize peak table}
\usage{
read.prosize.peaks(peaks.csv)
}
\arguments{
\item{peaks.csv}{A peak table CSV exported by ProSize.}
}
\value{
A list containing a list of assay metadata and a data frame of peaks (a partial \code{electrophoresis} object).
}
\description{
This function reads a peak table from the ProSize software saved in CSV format. The peak table must have been exported in the "alternate" format and it must include the "From" and "To" columns.
}
\seealso{
\code{\link{read.prosize}}, \code{\link{read.prosize.electropherogram}}, \code{\link{read.prosize.regions}}
}
