% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestation.R
\name{read.tapestation.csv}
\alias{read.tapestation.csv}
\title{Read a TapeStation CSV file}
\usage{
read.tapestation.csv(csv.file)
}
\arguments{
\item{csv.file}{The filename of an electropherogram CSV file exported from the TapeStation software. The file may be compressed with \code{gzip} and the filename is expected to end in \code{_Electropherogram.csv} or \code{_Electropherogram.csv.gz}. The filename can be a remote URL. The electropherogram is expected to be unaligned.}
}
\value{
A data frame with one row for each reading of each gel lane.
}
\description{
This function reads a CSV file of electrophoresis data exported from the TapeStation software.
}
\details{
Because the CSV file contains only the raw fluorescence data and not the metadata, this function is less useful by itself than when it is called inside \code{\link{read.tapestation}}.
}
\seealso{
\code{\link{read.tapestation}}, \code{\link{read.tapestation.xml}}, \code{\link{read.tapestation.gel.image}}
}
