% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestation.R
\name{read.tapestation}
\alias{read.tapestation}
\title{Read TapeStation data}
\usage{
read.tapestation(xml.file, gel.image.file = NULL, fit = "spline")
}
\arguments{
\item{xml.file}{The filename of an XML file exported from the TapeStation software. The filename is expected to end in \code{.xml} and the name before that extension is used as the name of the batch.}

\item{gel.image.file}{The filename of a TapeStation gel image with blue highlight, in PNG format. If \code{NULL}, the gel image file is expected to have the same name as the XML file with a different extension, e.g. \code{experiment1.xml} and \code{experiment1.png}, so if you name your files in that pattern you don't need to fill out this argument.}

\item{fit}{The method used to fit the mobility model of molecule length vs. migration distance, one of \code{"interpolation"} (linear interpolation via \code{\link{approxfun}}), \code{"spline"} (splines via \code{\link{splinefun}}), or \code{"regression"} (log-linear regression via \code{\link{lm}} with the model \code{relative.distance ~ log(length)}).}
}
\value{
An \code{electrophoresis} object containing the data from this TapeStation run.
}
\description{
This function reads TapeStation run data from an exported XML file and a PNG gel image, then fills out the results with estimates of molecule length, concentration, and molarity.
}
\details{
Spline fitting seems to perform reasonably well on all data. Agilent appears to use linear interpolation with DNA data and log-linear regression on RNA data, so you could choose those options if you want to reproduce the results of the software more precisely. However, linear interpolation creates sudden spikes in the derivative that make the concentration and molarity estimates unstable; spline fitting is basically a smoother version of that. Log-linear regression is the standard theoretical approach but does not actually fit the data very well; more sophisticated parametric models may be added in the future.
}
\seealso{
\code{\link{read.electrophoresis}}, \code{\link{read.bioanalyzer}}, \code{\link{read.tapestation.gel.image}}, \code{\link{read.tapestation.xml.file}}
}
