% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestation.R
\name{read.tapestation.xml}
\alias{read.tapestation.xml}
\title{Read a TapeStation XML file}
\usage{
read.tapestation.xml(xml.file)
}
\arguments{
\item{xml.file}{The filename of an XML file exported from the TapeStation software. The file may be compressed with \code{gzip} and the filename is expected to end in \code{.xml} or \code{.xml.gz}; the name before that extension is used as the name of the batch. The filename can be a remote URL.}
}
\value{
A list of some of the components of an \code{\link{electrophoresis}} object
}
\description{
This function reads an XML file exported from the TapeStation software.
}
\details{
Because the XML file contains only metadata and not the raw fluorescence data, this function is less useful by itself than when it is called inside \code{\link{read.tapestation}}.
}
\seealso{
\code{\link{read.tapestation}}, \code{\link{read.tapestation.gel.image}}
}
