% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{region.ratio}
\alias{region.ratio}
\title{Compare sums within regions}
\usage{
region.ratio(
  electrophoresis,
  ...,
  bound.variable = "length",
  sum.variable = "concentration"
)
}
\arguments{
\item{electrophoresis}{An \code{\link{electrophoresis}} object.}

\item{...}{Two or more pairs (vectors) of boundaries, e.g. \code{c(100, 200), c(200, 500), c(500, 700)}.}

\item{bound.variable}{Which variable the boundaries refer to.}

\item{sum.variable}{Which variable to sum in each region.}
}
\value{
A matrix of ratios of sums within the regions, each region relative to the first region, for each sample.
}
\description{
Given two or more regions (pairs of lower and upper bounds), calculate the ratio of the integrated sum of each additional region relative to the integrated sum of the first region.  The ratio is computed individually for each sample.
}
\seealso{
\code{\link{integrate.custom}}, \code{\link{dv200}}, \code{\link{illumina.library.ratio}}
}
