% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{scale.by.differential}
\alias{scale.by.differential}
\title{Scale data by a differential}
\usage{
\method{scale}{by.differential}(electrophoresis, x, y)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{x}{The name of the x-variable in \code{electrophoresis$data}.}

\item{y}{The name of the y-variable in \code{electrophoresis$data}.}
}
\value{
A vector of the y-values, one for each row of \code{electrophoresis$data}, divided by the differentials of the corresponding x-values.
}
\description{
Given an x-variable and a y-variable, this function scales the y-values from the observed data points by the differentials of the x-values. The resulting values of y/dx can then be used to make visually accurate graphs.
}
\seealso{
\code{\link{normalize.proportion}}
}
