% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{sparkline.electrophoresis}
\alias{sparkline.electrophoresis}
\title{Sparklines of electrophoresis data}
\usage{
sparkline.electrophoresis(
  ...,
  facets = sample.index ~ .,
  scales = "free_y",
  geom = "line",
  show.peaks = FALSE
)
}
\arguments{
\item{...}{Arguments passed to \code{\link{qplot.electrophoresis}}.}

\item{facets}{Passed to \code{\link{qplot.electrophoresis}} but the default is hardcoded for one column of sparklines.}

\item{scales, geom, show.peaks}{Passed to \code{\link{qplot.electrophoresis}} but the defaults are hardcoded for sparklines and probably do not make sense to change.}
}
\description{
This function is a shortcut that calls \code{\link{qplot.electrophoresis}} with customized settings to generate sparklines.
}
\details{
In addition to the hardcoded default arguments for \code{\link{qplot.electrophoresis}}, the following settings are added with \code{\link[ggplot2]{theme}}: \preformatted{
axis.text.y =       element_blank(),
axis.ticks.y =      element_blank(),
panel.grid =        element_blank(),
panel.background =  element_blank(),
strip.background =  element_blank(),
strip.text.y =      element_text(angle = 0, hjust = 0)
}
}
\references{
Tufte, Edward R. (1983) The Visual Display of Quantitative Information. Cheshire, Conn.: Graphics Press.
}
\seealso{
\code{\link{qplot.electrophoresis}}
}
