% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{subset.electrophoresis}
\alias{subset.electrophoresis}
\title{Subset samples an electrophoresis object}
\usage{
\method{subset}{electrophoresis}(electrophoresis, ...)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{...}{A logical expression indicating samples to keep, and any other arguments passed to \code{\link{subset}}.}
}
\value{
A new \code{electrophoresis} object containing only the data from the subset of samples that match the given expression, with sample indices renumbered and factors releveled.
}
\description{
This function takes a subset of the samples in an \code{electrophoresis} object.
}
