% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize.custom}
\alias{summarize.custom}
\title{Summarize lengths in a custom region}
\usage{
summarize.custom(electrophoresis, lower.bound = -Inf, upper.bound = Inf)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{lower.bound}{Lower boundary of the region to summarize.}

\item{upper.bound}{Upper boundary of the region to summarize.}
}
\description{
Compute summary statistics of the molecule length distribution between specified boundaries. The summary is computed individually for each sample.
}
\seealso{
\code{\link{summarize.peak}}, \code{\link{summarize.region}}
}
