% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize.peak.region}
\alias{summarize.peak.region}
\alias{summarize.peak}
\alias{summarize.region}
\title{Summarize lengths in peaks or regions}
\usage{
summarize.peak(electrophoresis, index = seq(nrow(electrophoresis$peaks)))

summarize.region(electrophoresis, index = seq(nrow(electrophoresis$regions)))
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{index}{The index, or a vector of indexes, of the peaks or regions to summarize (row numbers in \code{electrophoresis$peaks} or \code{electrophoresis$regions}).}
}
\description{
Compute summary statistics of the molecule length distribution in the reported peaks or regions in an \code{electrophoresis} object.
}
\seealso{
\code{\link{summarize.custom}}
}
