% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summarize.subset}
\alias{summarize.subset}
\title{Summarize lengths in a subset of data}
\usage{
summarize.subset(sample.frame)
}
\arguments{
\item{sample.frame}{A subset of an \code{\link{electrophoresis}} object containing only contiguous data from one sample.}
}
\description{
Compute summary statistics of the molecule length distribution in a subset of an \code{\link{electrophoresis}} object. This helper function is called by other functions that split the object into subsets.
}
\seealso{
\code{\link{summarize.peak.region}}, \code{\link{summarize.custom}}
}
