% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{violin.electrophoresis}
\alias{violin.electrophoresis}
\title{Violin plots of electrophoresis data}
\usage{
violin.electrophoresis(
  electrophoresis,
  x = "sample.name",
  y = "length",
  weight = "molarity",
  ...,
  normalize = T,
  include.ladder = FALSE,
  include.markers = FALSE,
  lower.marker.spread = 10,
  ylim = c(NA, NA),
  adjust = 0.05,
  title = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{electrophoresis}{An \code{\link{electrophoresis}} object.}

\item{x}{The name of the variable to use as the x-position of each violin as a character vector. Replicates with the same value of \code{x} will be grouped together with a single label. Probably \code{"sample.name"} unless you have added another sample annotation.}

\item{y}{The name of the variable to use as the y-axis of the violins, as a character vector. Usually one of \code{"time"}, \code{"aligned.time"}, \code{"distance"}, \code{"relative.distance"}, or \code{"length"}.}

\item{weight}{The name of the variable to use for the width of a violin at any given point on the y-axis, as a character vector. Usually one of \code{"fluorescence"}, \code{"concentration"}, or \code{"molarity"}.}

\item{...}{Additional aesthetics passed to \code{\link[ggplot2]{geom_violin}}, e.g. \code{fill}.}

\item{normalize}{Normalize the violins to the same area (within the y-limits), i.e. \code{scale = "area"} in \code{\link[ggplot2]{geom_violin}}. If false, their total areas will be proportional to their total non-negative weight values within the y-limits (\code{scale = "count"}). Alternatively, set to \code{NULL} and each violin (or group of violins with the same x-value) will have the same maximum width as the others (\code{scale = "width"}).}

\item{include.ladder}{If \code{FALSE}, graph only the actual samples and not the ladder well(s).}

\item{include.markers}{If \code{FALSE}, graph only data between the marker peaks.}

\item{lower.marker.spread}{If normalizing the totals or excluding marker peaks, extend the lower marker peak by this amount (via \code{\link{between.markers}}).}

\item{ylim}{Limits of y-axis.}

\item{adjust}{Bandwidth adjustment for the kernel smoothing in \code{\link[ggplot2]{geom_violin}}. Higher values will eliminate spikey noise but may also wipe out narrow peaks of interest, so this may require manual tuning depending on your data.}

\item{title, xlab, ylab}{Plot title, x-axis label, and y-axis label.}
}
\value{
A ggplot object containing several layers. You can draw it directly or customize it like any other ggplot object by adding more layers.
}
\description{
This function creates a violin plot from a \code{\link{electrophoresis}} object using \code{\link[ggplot2]{geom_violin}}.
}
\seealso{
\code{\link{ggplot.electrophoresis}}
}
