% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{electrophoresis}
\alias{electrophoresis}
\title{Electrophoresis class}
\usage{
electrophoresis(
  data = NULL,
  assay.info = NULL,
  samples = NULL,
  peaks = NULL,
  regions = NULL,
  mobility.functions = NULL
)
}
\description{
Electrophoresis class
}
