% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{get.x.name}
\alias{get.x.name}
\title{Get the original x-variable}
\usage{
get.x.name(electrophoresis, raw = FALSE, allow.multiple = FALSE)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{raw}{Whether to return the name of the raw variable instead of the aligned variable.}

\item{allow.multiple}{Whether to allow multiple raw variables (data combined from multiple platforms).}
}
\value{
A character giving the name of the x-variable.
}
\description{
This function takes an \code{electrophoresis} object and returns the name of the x-variable that was used to fit the mobility model.
}
\details{
If `raw == FALSE` the result should only be either \code{"aligned time"} for Bioanalyzer data or \code{"relative.distance"} for TapeStation data. If `raw == TRUE` the result should be \code{"time"} or \code{"distance"}.
}
