% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{integrate.custom}
\alias{integrate.custom}
\title{Integrate a variable in a custom region}
\usage{
integrate.custom(
  electrophoresis,
  lower.bound = -Inf,
  upper.bound = Inf,
  bound.variable = "length",
  sum.variable = "concentration"
)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{lower.bound}{Lower boundary of the region to integrate.}

\item{upper.bound}{Upper boundary of the region to integrate.}

\item{bound.variable}{Which variable the boundaries refer to.}

\item{sum.variable}{Which variable to sum in each region.}
}
\description{
Compute the sum of some electrophoresis variable in an \code{electrophoresis} object between specified boundaries. The variable is summed individually for each sample.
}
\seealso{
\code{\link{integrate.peaks}}, \code{\link{integrate.regions}}, \code{\link{region.ratio}}
}
