% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{read.electrophoresis}
\alias{read.electrophoresis}
\title{Read files into an electrophoresis object}
\usage{
read.electrophoresis(..., fit = "spline")
}
\arguments{
\item{...}{One or more XML files.}

\item{fit}{The method used to fit the mobility model of molecule length vs. migration distance, one of \code{"interpolation"} (linear interpolation via \code{\link{approxfun}}), \code{"spline"} (splines via \code{\link{splinefun}}), or \code{"regression"} (log-linear regression via \code{\link{lm}} with the model \code{relative.distance ~ log(length)}).}
}
\description{
This function reads one or more XML files exported from the Agilent software (and accompanying PNG files if from a TapeStation) and calls the appropriate function to read them into an \code{electrophoresis} object.
}
\details{
Multiple input files are combined with \code{\link{rbind.electrophoresis}}.

Warning: The filename of each TapeStation gel image must be the same as the corresponding XML file, e.g. \code{run1.png} goes with \code{run1.xml}.
}
\seealso{
\code{\link{read.bioanalyzer}}, \code{\link{read.tapestation}}
}
