% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioanalyzer.R, R/electrophoresis.R,
%   R/prosize.R, R/tapestation.R
\name{read.bioanalyzer}
\alias{read.bioanalyzer}
\alias{read.electrophoresis}
\alias{read.prosize}
\alias{read.prosize.zip}
\alias{read.tapestation}
\title{Read files into an electrophoresis object}
\usage{
read.bioanalyzer(xml.file, method = "hyman")

read.electrophoresis(
  ...,
  method = "hyman",
  mc.cores = if (.Platform$OS.type == "windows") 1 else detectCores()
)

read.prosize(
  electropherogram.csv,
  calibration.csv = NULL,
  peaks.csv = NULL,
  smear.csv = NULL,
  quality.csv = NULL,
  method = "hyman",
  batch = NULL
)

read.prosize.zip(prosize.zip, method = "hyman")

read.tapestation(
  xml.file,
  csv.file = NULL,
  gel.image.file = NULL,
  method = "hyman"
)
}
\arguments{
\item{xml.file}{The filename of an XML file exported from the Bioanalyzer or TapeStation software. The XML file may be compressed with `gzip` and the filename can be a remote URL. The filename is expected to end in \code{.xml} or \code{.xml.gz} and the name before that extension is used as the name of the batch. TapeStation XML files must have corresponding CSV or PNG files with matching names.}

\item{method}{The method used to fit the mobility model of molecule length vs. migration distance, either \code{"interpolation"} (linear interpolation via \code{\link{approxfun}}), \code{"loglinear"} (log-linear regression via \code{\link{lm}} with the model \code{relative.distance ~ log(length)}), or one of the methods of \code{\link{splinefun}} (monotone methods recommended). Can be abbreviated.}

\item{...}{One or more XML, CSV, or ZIP files exported from the Bioanalyzer, TapeStation, or ProSize software.}

\item{mc.cores}{Maximum number of CPU cores to use (passed to \code{\link[parallel]{mclapply}}). Only one core is used per input file.}

\item{electropherogram.csv}{An electropherogram CSV exported by ProSize.}

\item{calibration.csv}{A size calibration CSV exported by ProSize.}

\item{peaks.csv}{A peak table CSV exported by ProSize.}

\item{smear.csv}{A smear analysis CSV exported by ProSize.}

\item{quality.csv}{A quality table CSV exported by ProSize.}

\item{batch}{Name of the batch.}

\item{prosize.zip}{Path of a ZIP file containing the required CSVs exported by ProSize.}

\item{csv.file}{The filename of an electropherogram CSV file exported from the TapeStation software. The file may be compressed with \code{gzip} and the filename is expected to end in \code{_Electropherogram.csv} or \code{_Electropherogram.csv.gz}. The filename can be a remote URL. The electropherogram is expected to be unaligned.}

\item{gel.image.file}{The filename of a TapeStation gel image with blue highlight, in PNG format. The filename can be a URL.}
}
\description{
These functions read one or more XML, CSV, or ZIP files exported from the Agilent software (and accompanying unaligned electropherogram CSV or PNG files if from a TapeStation) and calls the appropriate function to read them into an \code{\link{electrophoresis}} object, which is filled out with estimates of molecule length, concentration, and molarity. \code{read.electrophoresis} is the easiest to use as it automatically infers the correct file type.
}
\details{
Spline fitting seems to perform reasonably well on all data. Agilent appears to use linear interpolation with DNA data and log-linear regression on RNA data, so you could choose those options if you want to reproduce the results of the software more precisely. However, linear interpolation creates sudden spikes in the derivative that make the concentration and molarity estimates unstable; spline fitting is basically a smoother version of that. Log-linear regression is the standard theoretical approach but does not actually fit the data very well; more sophisticated parametric models may be added in the future.
}
\section{Functions}{
\itemize{
\item \code{read.bioanalyzer}: Read a Bioanalyzer XML file

\item \code{read.prosize}: Read ProSize CSV files

\item \code{read.prosize.zip}: Read a ProSize ZIP file

\item \code{read.tapestation}: Read a TapeStation XML and PNG file pair
}}

