% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioanalyzer.R, R/electrophoresis.R,
%   R/tapestation.R
\name{read.bioanalyzer}
\alias{read.bioanalyzer}
\alias{read.electrophoresis}
\alias{read.tapestation}
\title{Read files into an electrophoresis object}
\usage{
read.bioanalyzer(xml.file, fit = "spline")

read.electrophoresis(..., fit = "spline")

read.tapestation(xml.file, gel.image.file = NULL, fit = "spline")
}
\arguments{
\item{xml.file}{The filename of an XML file exported from the Bioanalyzer or TapeStation software. The XML file may be compressed with `gzip`. The filename (without the file extension) is used as the name of the batch.}

\item{fit}{The method used to fit the mobility model of molecule length vs. migration distance, one of \code{"interpolation"} (linear interpolation via \code{\link{approxfun}}), \code{"spline"} (splines via \code{\link{splinefun}}), or \code{"regression"} (log-linear regression via \code{\link{lm}} with the model \code{relative.distance ~ log(length)}).}

\item{...}{One or more XML files exported from the Bioanalyzer or TapeStation software. TapeStation XML files must have corresponding PNG files with matching names.}

\item{gel.image.file}{The filename of a TapeStation gel image with blue highlight, in PNG format. If \code{NULL}, the gel image file is expected to have the same name as the XML file with a different extension, e.g. \code{experiment1.xml} and \code{experiment1.png}, so if you name your files in that pattern you don't need to fill out this argument.}
}
\description{
These functions read one or more XML files exported from the Agilent software (and accompanying PNG files if from a TapeStation) and calls the appropriate function to read them into an \code{electrophoresis} object, which is filled out with estimates of molecule length, concentration, and molarity. \code{read.electrophoresis} is the easiest to use as it automatically infers the correct file type.
}
\details{
Spline fitting seems to perform reasonably well on all data. Agilent appears to use linear interpolation with DNA data and log-linear regression on RNA data, so you could choose those options if you want to reproduce the results of the software more precisely. However, linear interpolation creates sudden spikes in the derivative that make the concentration and molarity estimates unstable; spline fitting is basically a smoother version of that. Log-linear regression is the standard theoretical approach but does not actually fit the data very well; more sophisticated parametric models may be added in the future.
}
\section{Functions}{
\itemize{
\item \code{read.bioanalyzer}: Read a Bioanalyzer XML file

\item \code{read.tapestation}: Read a TapeStation XML and PNG file pair
}}

