% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prosize.R
\name{read.prosize.peaks}
\alias{read.prosize.peaks}
\title{Read a ProSize peak table}
\usage{
read.prosize.peaks(csv.file)
}
\arguments{
\item{csv.file}{The filename of a peak table CSV exported by ProSize. The filename can be a URL.}
}
\value{
A list containing a list of assay metadata and a data frame of peaks (a partial \code{electrophoresis} object).
}
\description{
This function reads a peak table from the ProSize software saved in CSV format. The peak table must have been exported in the "alternate" format.
}
\seealso{
\code{\link{read.prosize}}, \code{\link{read.prosize.electropherogram}}, \code{\link{read.prosize.regions}}
}
