% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapestation.R
\name{read.tapestation.gel.image}
\alias{read.tapestation.gel.image}
\title{Read a TapeStation gel image}
\usage{
read.tapestation.gel.image(gel.image.file)
}
\arguments{
\item{gel.image.file}{The filename of a TapeStation gel image with blue highlight, in PNG format. The filename can be a URL.}
}
\value{
A data frame with one row for each vertical pixel of each gel lane.
}
\description{
This function reads a gel image exported from the TapeStation software and saved in PNG format. The gel image must include a blue highlight around one lane in order for the function to identify the boundaries of the gel area.
}
\details{
Because the gel image alone contains little metadata, this function returns only a simple data frame containing the fluorescence intensity vs. migration distance at every point in every lane of the gel (numbered from left to right). It is less useful by itself than when it is called inside \code{\link{read.tapestation}}.

Note: This function attempts to find the marker bands by their special color. But because this color is overlaid on the gel image, it is impossible to read the fluorescence intensity inside the markers, so it is reported as NA.
}
\seealso{
\code{\link{read.tapestation}}, \code{\link{read.tapestation.xml}}
}
