% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electrophoresis.R
\name{region.ratio}
\alias{region.ratio}
\title{Compare sums within regions}
\usage{
region.ratio(electrophoresis, bounds, ...)
}
\arguments{
\item{electrophoresis}{An \code{electrophoresis} object.}

\item{bounds}{A list of two or more pairs (vectors) of boundaries, e.g. \code{list(c(100, 200), c(200, 500), c(500, 700)}.}

\item{...}{Additional arguments passed to \code{\link{integrate.rawdata}}. In particular use \code{bound.variable} to set the boundary variable and \code{sum.variable} for the variable to sum.}
}
\value{
A matrix of ratios of sums within the regions, each region relative to the first region, for each sample.
}
\description{
Given two or more regions (pairs of lower and upper bounds), calculate the ratio of the integrated sum of each additional region relative to the integrated sum of the first region.  The ratio is compuate individually for each sample.
}
\seealso{
\code{\link{illumina.library.ratio}}
}
